requirements.txt
opencv-python
numpy==1.23.5
tensorflow-cpu==2.12.0
flask
pyttsx3
scikit-learn
matplotlib

